+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to Rainbow-CLI will be documented in this file.
 - CRRAINB-1986: Add command for adding developer role
 - CRRAINB-2095: Add options for retrieving application metrics by month and by year
 - CRRAINB-2106: Can't delete an application
+- CRRAINB-2114: Block and unblock an application (app_superadmin)
+- CRRAINB-2115: Deploy and dismiss an application (app_superadmin)
 
 ### [1.37.0] - 2018-25-02
 - RQRAINB-760: Rework mass-provisionning for devices management